Składanie zamówień
Elo

+++ src/main/java/com/JemToDobre/repository/ZamowieniaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ZamowieniaRepository extends JpaRepository<Zamowienia, Long> {
+public interface ZamowieniaRepository extends JpaRepository<Zamowienia, Integer> {
 }
